refactor: Move Logging functionality to support package
- Move deprecated log file configuration to dedicated support class
Cleans up the Extism API surface, as deprecated methods are now moved to Logging support class.
- Add log levels "off" and "error"
Rust defines the additional log levels "off" and "error".
